Refactor instrumentation code to separate file
